+++ src/__playground__/worker.ts
@@ -1,7 +1,7 @@
 import { define } from 'nanolith';
 
 export const worker = await define({
-    __initializeService() {
-        console.log('Initialized');
+    exit1() {
+        process.exit(1);
     },
 });